blender 批量传递UV

+++ blender_transfer_uv.py
@@ -5,16 +5,19 @@ import os
 transfer uvs
 """
 
-uv_exmaples_name = "smooth-0"
-model_path = "\\\\192.168.80.195\\data3\\LiyouWang\\smoothObjs"
-output_path = "\\\\192.168.80.195\\data3\\LiyouWang\\uv_objs"
+uv_exmaples_name = "polySurface1"
+model_path = "\\\\192.168.20.63\\ai\\Liyou_wang_data\\auto_blendshape\\blendshaps_56\\blendshapes"
+output_path = "\\\\192.168.20.63\\ai\\Liyou_wang_data\\auto_blendshape\\blendshaps_56\\uv_blendshape"
 filenames = os.listdir(model_path)
 for file in filenames:
-    bpy.ops.import_scene.obj(filepath=os.path.join(model_path, file))
-    bpy.data.objects[uv_exmaples_name].select=True
-    bpy.context.scene.objects.active=bpy.data.objects[uv_exmaples_name]
-    bpy.ops.object.join_uvs()
-    py.data.objects[uv_exmaples_name].select=False
-	bpy.ops.export_scene.obj(filepath=os.path.join(output_path, file), use_selection=True, use_materials=False)
-    bpy.ops.object.delete()
+    if file.endswith(".obj"):
+        bpy.ops.import_scene.obj(filepath=os.path.join(model_path, file))
+        bpy.data.objects[uv_exmaples_name].select=True
+        bpy.context.scene.objects.active=bpy.data.objects[uv_exmaples_name]
+        bpy.ops.object.join_uvs()
+        bpy.data.objects[uv_exmaples_name].select=False
+        bpy.ops.export_scene.obj(filepath=os.path.join(output_path, file), use_selection=True, use_materials=False)
+        bpy.ops.object.delete()
+    else:
+        print("skip {}".format(file))
